@@ -1,33 +1,41 @@
 {
     "private": true,
     "scripts": {
-        "dev": "npm run development",
-        "development": "cross-env NODE_ENV=development node_modules/webpack/bin/webpack.js --progress --hide-modules --config=node_modules/laravel-mix/setup/webpack.config.js",
-        "watch": "cross-env NODE_ENV=development node_modules/webpack/bin/webpack.js --watch --progress --hide-modules --config=node_modules/laravel-mix/setup/webpack.config.js",
-        "watch-poll": "npm run watch -- --watch-poll",
-        "hot": "cross-env NODE_ENV=development node_modules/webpack-dev-server/bin/webpack-dev-server.js --inline --hot --config=node_modules/laravel-mix/setup/webpack.config.js",
-        "prod": "npm run production",
-        "production": "cross-env NODE_ENV=production node_modules/webpack/bin/webpack.js --no-progress --hide-modules --config=node_modules/laravel-mix/setup/webpack.config.js"
+        "build": "webpack --hide-modules --mode=production",
+        "watch": "webpack --watch --hide-modules --mode=development"
     },
     "devDependencies": {
-        "@babel/plugin-proposal-class-properties": "^7.5.5",
-        "@babel/plugin-proposal-export-default-from": "^7.5.2",
-        "@babel/plugin-proposal-object-rest-spread": "^7.5.5",
-        "@babel/preset-env": "^7.5.5",
-        "@fortawesome/fontawesome-free": "^5.10.2",
-        "axios": "^0.18.0",
-        "backpack.css": "^2.0.0",
+        "@babel/core": "^7.4.5",
+        "@babel/plugin-proposal-class-properties": "^7.7.4",
+        "@babel/plugin-proposal-export-default-from": "^7.7.4",
+        "@babel/plugin-proposal-object-rest-spread": "^7.7.7",
+        "@babel/plugin-syntax-jsx": "^7.2.0",
+        "@babel/plugin-transform-object-assign": "^7.2.0",
+        "@babel/preset-env": "^7.7.7",
+        "@vue/babel-helper-vue-jsx-merge-props": "^1.0.0",
+        "@vue/babel-plugin-jsx": "^1.0.2",
+        "@vue/babel-plugin-transform-vue-jsx": "^1.1.2",
+        "@vue/compiler-sfc": "^3.0.5",
+        "babel-loader": "^8.0.6",
+        "copy-webpack-plugin": "^5.0.4",
+        "core-js": "^3.0.1",
         "cross-env": "^5.1",
-        "jquery": "^3.4.1",
-        "laravel-mix": "^4.0.7",
-        "nano-js": "https://github.com/vankizmann/nano-js.git#master",
-        "nano-ui": "https://github.com/vankizmann/nano-ui.git#master",
-        "raw-loader": "^1.0.0",
+        "css-loader": "^3.4.0",
+        "file-loader": "^4.1.0",
+        "ignore-loader": "^0.1.2",
+        "mini-css-extract-plugin": "^0.8.0",
+        "node-sass": "^4.12.0",
+        "postcss": "^8.1.1",
+        "postcss-inline-svg": "^4.1.0",
+        "postcss-loader": "^4.0.3",
         "sass": "^1.15.2",
         "sass-loader": "^7.1.0",
-        "vue": "^2.6.10",
-        "vue-router": "^3.0.2",
-        "vue-shortkey": "^3.1.6",
-        "vue-template-compiler": "^2.6.10"
+        "style-loader": "^1.1.2",
+        "terser-webpack-plugin": "^2.3.5",
+        "url-loader": "^2.1.0",
+        "vue": "^3.0.5",
+        "vue-loader": "^16.1.2",
+        "webpack": "^4.44.2",
+        "webpack-cli": "^3.3.12"
     }
 }
