@@ -74,7 +74,7 @@ export default {
         }
     },
 
-    destroyed()
+    unmounted()
     {
         Nano.Event.unbind('locale:changed', this._uid);
     },
