@@ -43,10 +43,11 @@
     </script>
 
 
-    {!! app('kyoto.assets')->style('google-font', '//fonts.googleapis.com/css2?family=Poppins:wght@400;500;600&display=swap') !!}
+    {!! app('kyoto.assets')->style('google-font', '//fonts.googleapis.com/css2?family=Inter:wght@400;700&display=swap') !!}
 
     {{-- {!! app('kyoto.assets')->script('vue', '//unpkg.com/vue@3') !!}
     {!! app('kyoto.assets')->script('vue-router', '//unpkg.com/vue-router@4') !!} --}}
+    {{-- <link rel="stylesheet" href="https://fonts.googleapis.com/css2?family=Roboto:wght@400;700&display=swap"> --}}
     {!! app('kyoto.assets')->script('script', 'kyoto/backend::js/index.js') !!}
     {!! app('kyoto.assets')->style('style', 'kyoto/backend::css/index.css') !!}
 
