@@ -1,22 +1,20 @@
-import Vue from 'vue';
-
 import KyoLanguages from "./pages/language/KyoLanguages";
-Vue.component(KyoLanguages.name, KyoLanguages);
+Kyoto.component(KyoLanguages.name, KyoLanguages);
 
 import KyoLanguageCreate from "./pages/language/KyoLanguageCreate";
-Vue.component(KyoLanguageCreate.name, KyoLanguageCreate);
+Kyoto.component(KyoLanguageCreate.name, KyoLanguageCreate);
 
 import KyoLanguageEdit from "./pages/language/KyoLanguageEdit";
-Vue.component(KyoLanguageEdit.name, KyoLanguageEdit);
+Kyoto.component(KyoLanguageEdit.name, KyoLanguageEdit);
 
 import KyoTranslations from "./pages/translation/KyoTranslations";
-Vue.component(KyoTranslations.name, KyoTranslations);
+Kyoto.component(KyoTranslations.name, KyoTranslations);
 
 import KyoTranslationCreate from "./pages/translation/KyoTranslationCreate";
-Vue.component(KyoTranslationCreate.name, KyoTranslationCreate);
+Kyoto.component(KyoTranslationCreate.name, KyoTranslationCreate);
 
 import KyoTranslationEdit from "./pages/translation/KyoTranslationEdit";
-Vue.component(KyoTranslationEdit.name, KyoTranslationEdit);
+Kyoto.component(KyoTranslationEdit.name, KyoTranslationEdit);
 
 if ( console && console.log ) {
     console.log('kyoto/language ready.');
