@@ -1,15 +1,15 @@
 <template>
-    <NLoader :visible="load" class="full-height-child">
+    <NLoader :visible="load" class="auto-height-child">
 
         <div class="grid grid--col">
 
             <KyoTitlebar class="col--flex-0-0" @delete="deleteItems">
 
-                <template slot="search">
+                <template v-slot:search>
                     <KyoTitlebarSearch v-model="query.search"></KyoTitlebarSearch>
                 </template>
 
-                <template slot="action">
+                <template v-slot:action>
                     <NButton type="primary" @click="$router.push({ name: 'KyoLanguageCreate' })">
                         {{ trans('Create language') }}
                     </NButton>
@@ -20,28 +20,16 @@
             <KyoDatatable class="col--flex-1-1" @row-dblclick="gotoEdit">
 
                 <NTableColumn type="string" prop="language" :label="trans('Language')" :fluid="true" :sort="true">
-                    <router-link slot-scope="{ value }" :to="{ name: 'KyoLanguageEdit', params: value }">{{ value.language }}</router-link>
+                    <template v-slot:default="{ item }" >
+                        <router-link :to="{ name: 'KyoLanguageEdit', params: item }">{{ item.language }}</router-link>
+                    </template>
                 </NTableColumn>
 
-                <NTableColumn type="string" prop="country" :label="trans('Country')" :fluid="true" :sort="true" :filter="true">
-                    <!-- Column -->
-                </NTableColumn>
-
-                <NTableColumn type="string" prop="locale" :label="trans('Locale')" :width="120" :sort="true" :filter="true">
-                    <!-- Column -->
-                </NTableColumn>
-
-                <NTableColumn type="string" prop="plate" :label="trans('Plate')" :width="120" :sort="true" :filter="true">
-                    <!-- Column -->
-                </NTableColumn>
-
-                <NTableColumn type="datetime" prop="updated_at" :label="trans('Modified')" :sort="true" :filter="true">
-                    <!-- Column -->
-                </NTableColumn>
-
-                <NTableColumn type="datetime" prop="created_at" :label="trans('Created')" :sort="true" :filter="true">
-                    <!-- Column -->
-                </NTableColumn>
+                <NTableColumn type="string" prop="country" :label="trans('Country')" :fluid="true" :sort="true" :filter="true"></NTableColumn>
+                <NTableColumn type="string" prop="locale" :label="trans('Locale')" :width="120" :sort="true" :filter="true"></NTableColumn>
+                <NTableColumn type="string" prop="plate" :label="trans('Plate')" :width="120" :sort="true" :filter="true"></NTableColumn>
+                <NTableColumn type="datetime" prop="updated_at" :label="trans('Modified')" :sort="true" :filter="true"></NTableColumn>
+                <NTableColumn type="datetime" prop="created_at" :label="trans('Created')" :sort="true" :filter="true"></NTableColumn>
 
             </KyoDatatable>
 
