@@ -1,15 +1,15 @@
 <template>
-    <NLoader :visible="load" class="full-height-child">
+    <NLoader :visible="load" class="auto-height-child">
 
         <div class="grid grid--col">
 
             <KyoTitlebar class="col--flex-0-0" @delete="deleteItems">
 
-                <template slot="search">
+                <template v-slot:search>
                     <KyoTitlebarSearch v-model="query.search"></KyoTitlebarSearch>
                 </template>
 
-                <template slot="action">
+                <template v-slot:action>
                     <NButton type="primary" @click="$router.push({ name: 'KyoTranslationCreate' })">
                         {{ trans('Create translation') }}
                     </NButton>
@@ -20,24 +20,15 @@
             <KyoDatatable class="col--flex-1-1" @row-dblclick="gotoEdit">
 
                 <NTableColumn type="string" prop="source" :label="trans('Source')" :fluid="true" :sort="true" :filter="true">
-                    <router-link slot-scope="{ value }" :to="{ name: 'KyoTranslationEdit', params: value }">{{ value.source }}</router-link>
+                    <template v-slot:default="{ item }">
+                        <router-link :to="{ name: 'KyoTranslationEdit', params: item }">{{ item.source }}</router-link>
+                    </template>
                 </NTableColumn>
 
-                <NTableColumn type="string" prop="target" :label="trans('Target')" :fluid="true" :sort="true" :filter="true">
-                    <!-- Column -->
-                </NTableColumn>
-
-                <NTableColumn type="string" prop="package" :label="trans('Package')" :sort="true" :filter="true">
-                    <!-- Column -->
-                </NTableColumn>
-
-                <NTableColumn type="datetime" prop="updated_at" :label="trans('Modified')" :sort="true" :filter="true">
-                    <!-- Column -->
-                </NTableColumn>
-
-                <NTableColumn type="datetime" prop="created_at" :label="trans('Created')" :sort="true" :filter="true">
-                    <!-- Column -->
-                </NTableColumn>
+                <NTableColumn type="string" prop="target" :label="trans('Target')" :fluid="true" :sort="true" :filter="true"></NTableColumn>
+                <NTableColumn type="string" prop="package" :label="trans('Package')" :sort="true" :filter="true"></NTableColumn>
+                <NTableColumn type="datetime" prop="updated_at" :label="trans('Modified')" :sort="true" :filter="true"></NTableColumn>
+                <NTableColumn type="datetime" prop="created_at" :label="trans('Created')" :sort="true" :filter="true"></NTableColumn>
 
             </KyoDatatable>
 
