@@ -3,7 +3,7 @@
         <div class="grid grid--col">
 
             <KyoTitlebar :back="true" class="col--flex-0-0" @delete="deleteItem">
-                <template slot="action">
+                <template v-slot:action>
                     <NButtonGroup>
                         <NButton type="primary" @click="updateItem">
                             {{ trans('Apply') }}
@@ -17,7 +17,7 @@
 
             <NForm :form="result" :errors="errors" class="kyo-dataform col--flex-1-0">
 
-                <NFormGroup icon="fa fa-info" :legend="trans('Info')">
+                <NFormGroup icon="fa fa-info" :label="trans('Info')">
                     <div class="grid grid-row grid--wrap grid--30">
 
                         <div class="col--1-1">
