@@ -86,27 +86,99 @@
 /************************************************************************/
 /******/ ({
 
-/***/ "./node_modules/babel-loader/lib/index.js?!./node_modules/vue-loader/dist/index.js?!./resources/js/pages/KyoPages2.vue?vue&type=script&lang=js":
-/*!***************************************************************************************************************************************************!*\
-  !*** ./node_modules/babel-loader/lib??ref--0!./node_modules/vue-loader/dist??ref--8-0!./resources/js/pages/KyoPages2.vue?vue&type=script&lang=js ***!
-  \***************************************************************************************************************************************************/
+/***/ "./node_modules/babel-loader/lib/index.js?!./node_modules/vue-loader/dist/index.js?!./resources/js/pages/KyoPageCreate.vue?vue&type=script&lang=js":
+/*!*******************************************************************************************************************************************************!*\
+  !*** ./node_modules/babel-loader/lib??ref--0!./node_modules/vue-loader/dist??ref--8-0!./resources/js/pages/KyoPageCreate.vue?vue&type=script&lang=js ***!
+  \*******************************************************************************************************************************************************/
 /*! exports provided: default */
 /***/ (function(module, __webpack_exports__, __webpack_require__) {
 
 "use strict";
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony default export */ __webpack_exports__[\"default\"] = ({\n  name: 'KyoPages',\n  localized: true,\n  urls: {\n    index: '/{locale}/kyoto/page/http/controllers/page/index',\n    \"delete\": '/{locale}/kyoto/page/http/controllers/page/delete'\n  },\n  defaults: function defaults() {\n    var query = {\n      page: 1,\n      limit: 25,\n      prop: 'updated_at',\n      dir: 'asc',\n      filter: [],\n      search: '',\n      columns: ['title', 'content']\n    };\n    return {\n      query: query\n    };\n  },\n  \"extends\": window.KyoIndex\n});\n\n//# sourceURL=webpack:///./resources/js/pages/KyoPages2.vue?./node_modules/babel-loader/lib??ref--0!./node_modules/vue-loader/dist??ref--8-0");
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony default export */ __webpack_exports__[\"default\"] = ({\n  name: 'KyoPageCreate',\n  localized: false,\n  urls: {\n    show: '/{locale}/kyoto/page/http/controllers/page/show',\n    store: '/{locale}/kyoto/page/http/controllers/page/store'\n  },\n  beforeMount: function beforeMount() {\n    this.$root.hideWebsite();\n  },\n  beforeDestroy: function beforeDestroy() {\n    this.$root.showWebsite();\n  },\n  methods: {\n    gotoEdit: function gotoEdit(row) {\n      this.$router.push({\n        name: 'KyoPageEdit',\n        params: row.data\n      });\n      console.log(row.data);\n    },\n    gotoIndex: function gotoIndex() {\n      this.$router.push({\n        name: 'KyoPages'\n      });\n    }\n  },\n  \"extends\": window.KyoForm\n});\n\n//# sourceURL=webpack:///./resources/js/pages/KyoPageCreate.vue?./node_modules/babel-loader/lib??ref--0!./node_modules/vue-loader/dist??ref--8-0");
 
 /***/ }),
 
-/***/ "./node_modules/vue-loader/dist/templateLoader.js?!./node_modules/vue-loader/dist/index.js?!./resources/js/pages/KyoPages2.vue?vue&type=template&id=7c630b54":
-/*!**************************************************************************************************************************************************************************!*\
-  !*** ./node_modules/vue-loader/dist/templateLoader.js??ref--5!./node_modules/vue-loader/dist??ref--8-0!./resources/js/pages/KyoPages2.vue?vue&type=template&id=7c630b54 ***!
-  \**************************************************************************************************************************************************************************/
+/***/ "./node_modules/babel-loader/lib/index.js?!./node_modules/vue-loader/dist/index.js?!./resources/js/pages/KyoPageEdit.vue?vue&type=script&lang=js":
+/*!*****************************************************************************************************************************************************!*\
+  !*** ./node_modules/babel-loader/lib??ref--0!./node_modules/vue-loader/dist??ref--8-0!./resources/js/pages/KyoPageEdit.vue?vue&type=script&lang=js ***!
+  \*****************************************************************************************************************************************************/
+/*! exports provided: default */
+/***/ (function(module, __webpack_exports__, __webpack_require__) {
+
+"use strict";
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony default export */ __webpack_exports__[\"default\"] = ({\n  name: 'KyoPageEdit',\n  localized: true,\n  urls: {\n    show: '/{locale}/kyoto/page/http/controllers/page/show',\n    update: '/{locale}/kyoto/page/http/controllers/page/update',\n    \"delete\": '/{locale}/kyoto/page/http/controllers/page/delete'\n  },\n  beforeMount: function beforeMount() {// this.$root.hideWebsite();\n  },\n  beforeDestroy: function beforeDestroy() {// this.$root.showWebsite();\n  },\n  methods: {\n    gotoIndex: function gotoIndex() {\n      this.$router.push({\n        name: 'KyoPages'\n      });\n    }\n  },\n  \"extends\": window.KyoForm\n});\n\n//# sourceURL=webpack:///./resources/js/pages/KyoPageEdit.vue?./node_modules/babel-loader/lib??ref--0!./node_modules/vue-loader/dist??ref--8-0");
+
+/***/ }),
+
+/***/ "./node_modules/babel-loader/lib/index.js?!./node_modules/vue-loader/dist/index.js?!./resources/js/pages/KyoPages.vue?vue&type=script&lang=js":
+/*!**************************************************************************************************************************************************!*\
+  !*** ./node_modules/babel-loader/lib??ref--0!./node_modules/vue-loader/dist??ref--8-0!./resources/js/pages/KyoPages.vue?vue&type=script&lang=js ***!
+  \**************************************************************************************************************************************************/
+/*! exports provided: default */
+/***/ (function(module, __webpack_exports__, __webpack_require__) {
+
+"use strict";
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony default export */ __webpack_exports__[\"default\"] = ({\n  name: 'KyoPages',\n  localized: true,\n  urls: {\n    index: '/{locale}/kyoto/page/http/controllers/page/index',\n    \"delete\": '/{locale}/kyoto/page/http/controllers/page/delete'\n  },\n  defaults: function defaults() {\n    var query = {\n      page: 1,\n      limit: 25,\n      prop: 'updated_at',\n      dir: 'asc',\n      filter: [],\n      search: '',\n      columns: ['title', 'content']\n    };\n    return {\n      query: query\n    };\n  },\n  \"extends\": window.KyoIndex\n});\n\n//# sourceURL=webpack:///./resources/js/pages/KyoPages.vue?./node_modules/babel-loader/lib??ref--0!./node_modules/vue-loader/dist??ref--8-0");
+
+/***/ }),
+
+/***/ "./node_modules/babel-loader/lib/index.js?!./node_modules/vue-loader/dist/index.js?!./resources/js/widgets/KyoWidgetHeadline.vue?vue&type=script&lang=js":
+/*!*************************************************************************************************************************************************************!*\
+  !*** ./node_modules/babel-loader/lib??ref--0!./node_modules/vue-loader/dist??ref--8-0!./resources/js/widgets/KyoWidgetHeadline.vue?vue&type=script&lang=js ***!
+  \*************************************************************************************************************************************************************/
+/*! exports provided: default */
+/***/ (function(module, __webpack_exports__, __webpack_require__) {
+
+"use strict";
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony default export */ __webpack_exports__[\"default\"] = ({\n  name: 'KyoWidgetHeadline',\n  inject: {\n    NDraggableItem: {\n      \"default\": undefined\n    }\n  },\n  computed: {\n    value: function value() {\n      return this.NDraggableItem.veItem.value;\n    }\n  },\n  mounted: function mounted() {\n    this.$watch('value', Nano.Any.debounce(this.NDraggableItem[\"export\"], 100), {\n      deep: true\n    });\n  }\n});\n\n//# sourceURL=webpack:///./resources/js/widgets/KyoWidgetHeadline.vue?./node_modules/babel-loader/lib??ref--0!./node_modules/vue-loader/dist??ref--8-0");
+
+/***/ }),
+
+/***/ "./node_modules/vue-loader/dist/templateLoader.js?!./node_modules/vue-loader/dist/index.js?!./resources/js/pages/KyoPageCreate.vue?vue&type=template&id=22f38b53":
+/*!******************************************************************************************************************************************************************************!*\
+  !*** ./node_modules/vue-loader/dist/templateLoader.js??ref--5!./node_modules/vue-loader/dist??ref--8-0!./resources/js/pages/KyoPageCreate.vue?vue&type=template&id=22f38b53 ***!
+  \******************************************************************************************************************************************************************************/
 /*! exports provided: render */
 /***/ (function(module, __webpack_exports__, __webpack_require__) {
 
 "use strict";
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"render\", function() { return render; });\n/* harmony import */ var vue__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! vue */ \"vue\");\n/* harmony import */ var vue__WEBPACK_IMPORTED_MODULE_0___default = /*#__PURE__*/__webpack_require__.n(vue__WEBPACK_IMPORTED_MODULE_0__);\n\n\nconst _hoisted_1 = { class: \"grid grid--col\" }\n\nfunction render(_ctx, _cache, $props, $setup, $data, $options) {\n  const _component_KyoTitlebarSearch = Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"resolveComponent\"])(\"KyoTitlebarSearch\")\n  const _component_NButton = Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"resolveComponent\"])(\"NButton\")\n  const _component_KyoTitlebar = Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"resolveComponent\"])(\"KyoTitlebar\")\n  const _component_NTableColumn = Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"resolveComponent\"])(\"NTableColumn\")\n  const _component_router_link = Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"resolveComponent\"])(\"router-link\")\n  const _component_KyoDatatable = Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"resolveComponent\"])(\"KyoDatatable\")\n  const _component_NLoader = Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"resolveComponent\"])(\"NLoader\")\n\n  return (Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"openBlock\"])(), Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createBlock\"])(_component_NLoader, {\n    visible: _ctx.load,\n    class: \"full-height-child\"\n  }, {\n    default: Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"withCtx\"])(() => [\n      Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createVNode\"])(\"div\", _hoisted_1, [\n        Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createVNode\"])(_component_KyoTitlebar, {\n          class: \"col--flex-0-0\",\n          onDelete: _ctx.deleteItems\n        }, {\n          search: Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"withCtx\"])(() => [\n            Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createVNode\"])(_component_KyoTitlebarSearch, {\n              modelValue: _ctx.query.search,\n              \"onUpdate:modelValue\": _cache[1] || (_cache[1] = $event => (_ctx.query.search = $event))\n            }, null, 8 /* PROPS */, [\"modelValue\"])\n          ]),\n          action: Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"withCtx\"])(() => [\n            Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createVNode\"])(_component_NButton, {\n              type: \"primary\",\n              onClick: _cache[2] || (_cache[2] = $event => (_ctx.$router.push({ name: 'KyoPageCreate' })))\n            }, {\n              default: Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"withCtx\"])(() => [\n                Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createTextVNode\"])(Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"toDisplayString\"])(_ctx.trans('Create page')), 1 /* TEXT */)\n              ]),\n              _: 1 /* STABLE */\n            })\n          ]),\n          _: 1 /* STABLE */\n        }, 8 /* PROPS */, [\"onDelete\"]),\n        Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createVNode\"])(_component_KyoDatatable, {\n          class: \"col--flex-1-1\",\n          onRowDblclick: _ctx.gotoEdit,\n          group: ['page', 'transaction']\n        }, {\n          default: Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"withCtx\"])(() => [\n            Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createVNode\"])(_component_NTableColumn, {\n              type: \"boolean\",\n              prop: \"state\",\n              label: _ctx.trans('State'),\n              sort: true,\n              filter: true\n            }, {\n              default: Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"withCtx\"])(() => [\n                Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createCommentVNode\"])(\" Column \")\n              ]),\n              _: 1 /* STABLE */\n            }, 8 /* PROPS */, [\"label\"]),\n            Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createVNode\"])(_component_NTableColumn, {\n              type: \"boolean\",\n              prop: \"hide\",\n              label: _ctx.trans('Hide'),\n              sort: true,\n              filter: true\n            }, {\n              default: Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"withCtx\"])(() => [\n                Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createCommentVNode\"])(\" Column \")\n              ]),\n              _: 1 /* STABLE */\n            }, 8 /* PROPS */, [\"label\"]),\n            Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createVNode\"])(_component_NTableColumn, {\n              type: \"string\",\n              prop: \"title\",\n              label: _ctx.trans('Title'),\n              fluid: true,\n              sort: true,\n              filter: true\n            }, {\n              default: Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"withCtx\"])(() => [\n                Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createVNode\"])(_component_router_link, {\n                  to: { name: 'KyoPageEdit', params: _ctx.item }\n                }, {\n                  default: Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"withCtx\"])((item) => [\n                    Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createTextVNode\"])(Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"toDisplayString\"])(item.title), 1 /* TEXT */)\n                  ]),\n                  _: 1 /* STABLE */\n                }, 8 /* PROPS */, [\"to\"])\n              ]),\n              _: 1 /* STABLE */\n            }, 8 /* PROPS */, [\"label\"]),\n            Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createVNode\"])(_component_NTableColumn, {\n              type: \"datetime\",\n              prop: \"updated_at\",\n              label: _ctx.trans('Modified'),\n              sort: true,\n              filter: true\n            }, {\n              default: Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"withCtx\"])(() => [\n                Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createCommentVNode\"])(\" Column \")\n              ]),\n              _: 1 /* STABLE */\n            }, 8 /* PROPS */, [\"label\"]),\n            Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createVNode\"])(_component_NTableColumn, {\n              type: \"datetime\",\n              prop: \"created_at\",\n              label: _ctx.trans('Created'),\n              sort: true,\n              filter: true\n            }, {\n              default: Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"withCtx\"])(() => [\n                Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createCommentVNode\"])(\" Column \")\n              ]),\n              _: 1 /* STABLE */\n            }, 8 /* PROPS */, [\"label\"])\n          ]),\n          _: 1 /* STABLE */\n        }, 8 /* PROPS */, [\"onRowDblclick\"])\n      ])\n    ]),\n    _: 1 /* STABLE */\n  }, 8 /* PROPS */, [\"visible\"]))\n}\n\n//# sourceURL=webpack:///./resources/js/pages/KyoPages2.vue?./node_modules/vue-loader/dist/templateLoader.js??ref--5!./node_modules/vue-loader/dist??ref--8-0");
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"render\", function() { return render; });\n/* harmony import */ var vue__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! vue */ \"vue\");\n/* harmony import */ var vue__WEBPACK_IMPORTED_MODULE_0___default = /*#__PURE__*/__webpack_require__.n(vue__WEBPACK_IMPORTED_MODULE_0__);\n\n\nconst _hoisted_1 = { class: \"grid grid--col\" }\n\nfunction render(_ctx, _cache, $props, $setup, $data, $options) {\n  const _component_NButton = Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"resolveComponent\"])(\"NButton\")\n  const _component_NButtonGroup = Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"resolveComponent\"])(\"NButtonGroup\")\n  const _component_KyoTitlebar = Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"resolveComponent\"])(\"KyoTitlebar\")\n  const _component_NSwitch = Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"resolveComponent\"])(\"NSwitch\")\n  const _component_NFormItem = Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"resolveComponent\"])(\"NFormItem\")\n  const _component_NFormGroup = Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"resolveComponent\"])(\"NFormGroup\")\n  const _component_NInput = Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"resolveComponent\"])(\"NInput\")\n  const _component_NTextarea = Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"resolveComponent\"])(\"NTextarea\")\n  const _component_NForm = Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"resolveComponent\"])(\"NForm\")\n  const _component_KyoPageWidgets = Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"resolveComponent\"])(\"KyoPageWidgets\")\n  const _component_NLoader = Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"resolveComponent\"])(\"NLoader\")\n\n  return (Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"openBlock\"])(), Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createBlock\"])(_component_NLoader, {\n    visible: _ctx.load,\n    class: \"full-height-child\"\n  }, {\n    default: Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"withCtx\"])(() => [\n      Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createVNode\"])(\"div\", _hoisted_1, [\n        Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createVNode\"])(_component_KyoTitlebar, {\n          link: { name: 'KyoPages' },\n          class: \"col--flex-0-0\"\n        }, {\n          action: Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"withCtx\"])(() => [\n            Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createVNode\"])(_component_NButtonGroup, null, {\n              default: Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"withCtx\"])(() => [\n                Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createVNode\"])(_component_NButton, {\n                  type: \"primary\",\n                  onClick: _ctx.storeItem\n                }, {\n                  default: Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"withCtx\"])(() => [\n                    Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createTextVNode\"])(Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"toDisplayString\"])(_ctx.trans('Apply')), 1 /* TEXT */)\n                  ]),\n                  _: 1 /* STABLE */\n                }, 8 /* PROPS */, [\"onClick\"]),\n                Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createVNode\"])(_component_NButton, {\n                  type: \"primary\",\n                  onClick: _ctx.storeItemClose\n                }, {\n                  default: Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"withCtx\"])(() => [\n                    Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createTextVNode\"])(Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"toDisplayString\"])(_ctx.trans('Save')), 1 /* TEXT */)\n                  ]),\n                  _: 1 /* STABLE */\n                }, 8 /* PROPS */, [\"onClick\"])\n              ]),\n              _: 1 /* STABLE */\n            })\n          ]),\n          _: 1 /* STABLE */\n        }),\n        Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createVNode\"])(_component_KyoPageWidgets, {\n          class: \"col--flex-1-0\",\n          modelValue: _ctx.result.builder,\n          \"onUpdate:modelValue\": _cache[6] || (_cache[6] = $event => (_ctx.result.builder = $event))\n        }, {\n          default: Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"withCtx\"])(() => [\n            Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createVNode\"])(_component_NForm, {\n              form: _ctx.result,\n              errors: _ctx.errors\n            }, {\n              default: Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"withCtx\"])(() => [\n                Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createVNode\"])(_component_NFormGroup, {\n                  icon: \"fa fa-cog\",\n                  label: _ctx.trans('Settings')\n                }, {\n                  default: Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"withCtx\"])(() => [\n                    Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createVNode\"])(_component_NFormItem, {\n                      label: _ctx.trans('State'),\n                      prop: \"state\"\n                    }, {\n                      default: Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"withCtx\"])(() => [\n                        Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createVNode\"])(_component_NSwitch, {\n                          modelValue: _ctx.result.state,\n                          \"onUpdate:modelValue\": _cache[1] || (_cache[1] = $event => (_ctx.result.state = $event)),\n                          \"on-value\": 1,\n                          \"off-value\": 0\n                        }, {\n                          default: Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"withCtx\"])(() => [\n                            Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createTextVNode\"])(Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"toDisplayString\"])(_ctx.trans('Item is enabled')), 1 /* TEXT */)\n                          ]),\n                          _: 1 /* STABLE */\n                        }, 8 /* PROPS */, [\"modelValue\"])\n                      ]),\n                      _: 1 /* STABLE */\n                    }, 8 /* PROPS */, [\"label\"]),\n                    Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createVNode\"])(_component_NFormItem, {\n                      label: _ctx.trans('Hide'),\n                      prop: \"hide\"\n                    }, {\n                      default: Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"withCtx\"])(() => [\n                        Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createVNode\"])(_component_NSwitch, {\n                          modelValue: _ctx.result.hide,\n                          \"onUpdate:modelValue\": _cache[2] || (_cache[2] = $event => (_ctx.result.hide = $event)),\n                          \"on-value\": 1,\n                          \"off-value\": 0\n                        }, {\n                          default: Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"withCtx\"])(() => [\n                            Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createTextVNode\"])(Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"toDisplayString\"])(_ctx.trans('Item is hidden')), 1 /* TEXT */)\n                          ]),\n                          _: 1 /* STABLE */\n                        }, 8 /* PROPS */, [\"modelValue\"])\n                      ]),\n                      _: 1 /* STABLE */\n                    }, 8 /* PROPS */, [\"label\"])\n                  ]),\n                  _: 1 /* STABLE */\n                }, 8 /* PROPS */, [\"label\"]),\n                Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createVNode\"])(_component_NFormGroup, {\n                  icon: \"fa fa-compass\",\n                  label: _ctx.trans('Menu')\n                }, {\n                  default: Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"withCtx\"])(() => [\n                    Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createVNode\"])(_component_NFormItem, {\n                      label: _ctx.trans('Title'),\n                      prop: \"title\"\n                    }, {\n                      default: Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"withCtx\"])(() => [\n                        Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createVNode\"])(_component_NInput, {\n                          modelValue: _ctx.result.title,\n                          \"onUpdate:modelValue\": _cache[3] || (_cache[3] = $event => (_ctx.result.title = $event))\n                        }, null, 8 /* PROPS */, [\"modelValue\"])\n                      ]),\n                      _: 1 /* STABLE */\n                    }, 8 /* PROPS */, [\"label\"]),\n                    Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createVNode\"])(_component_NFormItem, {\n                      label: _ctx.trans('Slug'),\n                      prop: \"slug\"\n                    }, {\n                      default: Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"withCtx\"])(() => [\n                        Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createVNode\"])(_component_NInput, {\n                          modelValue: _ctx.result.slug,\n                          \"onUpdate:modelValue\": _cache[4] || (_cache[4] = $event => (_ctx.result.slug = $event))\n                        }, null, 8 /* PROPS */, [\"modelValue\"])\n                      ]),\n                      _: 1 /* STABLE */\n                    }, 8 /* PROPS */, [\"label\"]),\n                    Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createVNode\"])(_component_NFormItem, {\n                      label: _ctx.trans('Content'),\n                      prop: \"content\"\n                    }, {\n                      default: Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"withCtx\"])(() => [\n                        Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createVNode\"])(_component_NTextarea, {\n                          modelValue: _ctx.result.content,\n                          \"onUpdate:modelValue\": _cache[5] || (_cache[5] = $event => (_ctx.result.content = $event))\n                        }, null, 8 /* PROPS */, [\"modelValue\"])\n                      ]),\n                      _: 1 /* STABLE */\n                    }, 8 /* PROPS */, [\"label\"])\n                  ]),\n                  _: 1 /* STABLE */\n                }, 8 /* PROPS */, [\"label\"])\n              ]),\n              _: 1 /* STABLE */\n            }, 8 /* PROPS */, [\"form\", \"errors\"])\n          ]),\n          _: 1 /* STABLE */\n        }, 8 /* PROPS */, [\"modelValue\"])\n      ])\n    ]),\n    _: 1 /* STABLE */\n  }, 8 /* PROPS */, [\"visible\"]))\n}\n\n//# sourceURL=webpack:///./resources/js/pages/KyoPageCreate.vue?./node_modules/vue-loader/dist/templateLoader.js??ref--5!./node_modules/vue-loader/dist??ref--8-0");
+
+/***/ }),
+
+/***/ "./node_modules/vue-loader/dist/templateLoader.js?!./node_modules/vue-loader/dist/index.js?!./resources/js/pages/KyoPageEdit.vue?vue&type=template&id=d1e6f5be":
+/*!****************************************************************************************************************************************************************************!*\
+  !*** ./node_modules/vue-loader/dist/templateLoader.js??ref--5!./node_modules/vue-loader/dist??ref--8-0!./resources/js/pages/KyoPageEdit.vue?vue&type=template&id=d1e6f5be ***!
+  \****************************************************************************************************************************************************************************/
+/*! exports provided: render */
+/***/ (function(module, __webpack_exports__, __webpack_require__) {
+
+"use strict";
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"render\", function() { return render; });\n/* harmony import */ var vue__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! vue */ \"vue\");\n/* harmony import */ var vue__WEBPACK_IMPORTED_MODULE_0___default = /*#__PURE__*/__webpack_require__.n(vue__WEBPACK_IMPORTED_MODULE_0__);\n\n\nconst _hoisted_1 = { class: \"grid grid--col\" }\n\nfunction render(_ctx, _cache, $props, $setup, $data, $options) {\n  const _component_NButton = Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"resolveComponent\"])(\"NButton\")\n  const _component_NButtonGroup = Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"resolveComponent\"])(\"NButtonGroup\")\n  const _component_KyoTitlebar = Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"resolveComponent\"])(\"KyoTitlebar\")\n  const _component_NInput = Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"resolveComponent\"])(\"NInput\")\n  const _component_NFormItem = Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"resolveComponent\"])(\"NFormItem\")\n  const _component_NTextarea = Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"resolveComponent\"])(\"NTextarea\")\n  const _component_NFormGroup = Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"resolveComponent\"])(\"NFormGroup\")\n  const _component_NSwitch = Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"resolveComponent\"])(\"NSwitch\")\n  const _component_NForm = Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"resolveComponent\"])(\"NForm\")\n  const _component_KyoPageWidgets = Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"resolveComponent\"])(\"KyoPageWidgets\")\n  const _component_NLoader = Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"resolveComponent\"])(\"NLoader\")\n\n  return (Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"openBlock\"])(), Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createBlock\"])(_component_NLoader, {\n    visible: _ctx.load,\n    class: \"full-height-child\"\n  }, {\n    default: Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"withCtx\"])(() => [\n      Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createVNode\"])(\"div\", _hoisted_1, [\n        Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createVNode\"])(_component_KyoTitlebar, {\n          link: { name: 'KyoPages' },\n          class: \"col--flex-0-0\",\n          onDelete: _ctx.deleteItem\n        }, {\n          action: Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"withCtx\"])(() => [\n            Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createVNode\"])(_component_NButtonGroup, null, {\n              default: Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"withCtx\"])(() => [\n                Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createVNode\"])(_component_NButton, {\n                  type: \"primary\",\n                  onClick: _ctx.updateItem\n                }, {\n                  default: Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"withCtx\"])(() => [\n                    Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createTextVNode\"])(Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"toDisplayString\"])(_ctx.trans('Apply')), 1 /* TEXT */)\n                  ]),\n                  _: 1 /* STABLE */\n                }, 8 /* PROPS */, [\"onClick\"]),\n                Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createVNode\"])(_component_NButton, {\n                  type: \"primary\",\n                  onClick: _ctx.updateCloseItem\n                }, {\n                  default: Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"withCtx\"])(() => [\n                    Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createTextVNode\"])(Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"toDisplayString\"])(_ctx.trans('Save')), 1 /* TEXT */)\n                  ]),\n                  _: 1 /* STABLE */\n                }, 8 /* PROPS */, [\"onClick\"])\n              ]),\n              _: 1 /* STABLE */\n            })\n          ]),\n          _: 1 /* STABLE */\n        }, 8 /* PROPS */, [\"onDelete\"]),\n        Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createVNode\"])(_component_KyoPageWidgets, {\n          class: \"col--flex-1-0\",\n          modelValue: _ctx.result.builder,\n          \"onUpdate:modelValue\": _cache[6] || (_cache[6] = $event => (_ctx.result.builder = $event))\n        }, {\n          default: Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"withCtx\"])(() => [\n            Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createVNode\"])(_component_NForm, {\n              form: _ctx.result,\n              errors: _ctx.errors\n            }, {\n              default: Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"withCtx\"])(() => [\n                Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createVNode\"])(_component_NFormGroup, {\n                  icon: \"fa fa-compass\",\n                  label: _ctx.trans('Menu')\n                }, {\n                  default: Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"withCtx\"])(() => [\n                    Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createVNode\"])(_component_NFormItem, {\n                      label: _ctx.trans('Title'),\n                      prop: \"title\"\n                    }, {\n                      default: Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"withCtx\"])(() => [\n                        Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createVNode\"])(_component_NInput, {\n                          modelValue: _ctx.result.title,\n                          \"onUpdate:modelValue\": _cache[1] || (_cache[1] = $event => (_ctx.result.title = $event))\n                        }, null, 8 /* PROPS */, [\"modelValue\"])\n                      ]),\n                      _: 1 /* STABLE */\n                    }, 8 /* PROPS */, [\"label\"]),\n                    Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createVNode\"])(_component_NFormItem, {\n                      label: _ctx.trans('Slug'),\n                      prop: \"slug\"\n                    }, {\n                      default: Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"withCtx\"])(() => [\n                        Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createVNode\"])(_component_NInput, {\n                          modelValue: _ctx.result.slug,\n                          \"onUpdate:modelValue\": _cache[2] || (_cache[2] = $event => (_ctx.result.slug = $event))\n                        }, null, 8 /* PROPS */, [\"modelValue\"])\n                      ]),\n                      _: 1 /* STABLE */\n                    }, 8 /* PROPS */, [\"label\"]),\n                    Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createVNode\"])(_component_NFormItem, {\n                      label: _ctx.trans('Content'),\n                      prop: \"content\"\n                    }, {\n                      default: Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"withCtx\"])(() => [\n                        Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createVNode\"])(_component_NTextarea, {\n                          modelValue: _ctx.result.content,\n                          \"onUpdate:modelValue\": _cache[3] || (_cache[3] = $event => (_ctx.result.content = $event))\n                        }, null, 8 /* PROPS */, [\"modelValue\"])\n                      ]),\n                      _: 1 /* STABLE */\n                    }, 8 /* PROPS */, [\"label\"])\n                  ]),\n                  _: 1 /* STABLE */\n                }, 8 /* PROPS */, [\"label\"]),\n                Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createVNode\"])(_component_NFormGroup, {\n                  icon: \"fa fa-cog\",\n                  label: _ctx.trans('Settings')\n                }, {\n                  default: Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"withCtx\"])(() => [\n                    Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createVNode\"])(_component_NFormItem, {\n                      label: _ctx.trans('State'),\n                      prop: \"state\"\n                    }, {\n                      default: Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"withCtx\"])(() => [\n                        Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createVNode\"])(_component_NSwitch, {\n                          modelValue: _ctx.result.state,\n                          \"onUpdate:modelValue\": _cache[4] || (_cache[4] = $event => (_ctx.result.state = $event)),\n                          \"on-value\": 1,\n                          \"off-value\": 0\n                        }, {\n                          default: Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"withCtx\"])(() => [\n                            Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createTextVNode\"])(Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"toDisplayString\"])(_ctx.trans('Item is enabled')), 1 /* TEXT */)\n                          ]),\n                          _: 1 /* STABLE */\n                        }, 8 /* PROPS */, [\"modelValue\"])\n                      ]),\n                      _: 1 /* STABLE */\n                    }, 8 /* PROPS */, [\"label\"]),\n                    Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createVNode\"])(_component_NFormItem, {\n                      label: _ctx.trans('Hide'),\n                      prop: \"hide\"\n                    }, {\n                      default: Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"withCtx\"])(() => [\n                        Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createVNode\"])(_component_NSwitch, {\n                          modelValue: _ctx.result.hide,\n                          \"onUpdate:modelValue\": _cache[5] || (_cache[5] = $event => (_ctx.result.hide = $event)),\n                          \"on-value\": 1,\n                          \"off-value\": 0\n                        }, {\n                          default: Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"withCtx\"])(() => [\n                            Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createTextVNode\"])(Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"toDisplayString\"])(_ctx.trans('Item is hidden')), 1 /* TEXT */)\n                          ]),\n                          _: 1 /* STABLE */\n                        }, 8 /* PROPS */, [\"modelValue\"])\n                      ]),\n                      _: 1 /* STABLE */\n                    }, 8 /* PROPS */, [\"label\"])\n                  ]),\n                  _: 1 /* STABLE */\n                }, 8 /* PROPS */, [\"label\"])\n              ]),\n              _: 1 /* STABLE */\n            }, 8 /* PROPS */, [\"form\", \"errors\"])\n          ]),\n          _: 1 /* STABLE */\n        }, 8 /* PROPS */, [\"modelValue\"])\n      ])\n    ]),\n    _: 1 /* STABLE */\n  }, 8 /* PROPS */, [\"visible\"]))\n}\n\n//# sourceURL=webpack:///./resources/js/pages/KyoPageEdit.vue?./node_modules/vue-loader/dist/templateLoader.js??ref--5!./node_modules/vue-loader/dist??ref--8-0");
+
+/***/ }),
+
+/***/ "./node_modules/vue-loader/dist/templateLoader.js?!./node_modules/vue-loader/dist/index.js?!./resources/js/pages/KyoPages.vue?vue&type=template&id=671d7ee8":
+/*!*************************************************************************************************************************************************************************!*\
+  !*** ./node_modules/vue-loader/dist/templateLoader.js??ref--5!./node_modules/vue-loader/dist??ref--8-0!./resources/js/pages/KyoPages.vue?vue&type=template&id=671d7ee8 ***!
+  \*************************************************************************************************************************************************************************/
+/*! exports provided: render */
+/***/ (function(module, __webpack_exports__, __webpack_require__) {
+
+"use strict";
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"render\", function() { return render; });\n/* harmony import */ var vue__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! vue */ \"vue\");\n/* harmony import */ var vue__WEBPACK_IMPORTED_MODULE_0___default = /*#__PURE__*/__webpack_require__.n(vue__WEBPACK_IMPORTED_MODULE_0__);\n\n\nconst _hoisted_1 = { class: \"grid grid--col\" }\n\nfunction render(_ctx, _cache, $props, $setup, $data, $options) {\n  const _component_KyoTitlebarSearch = Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"resolveComponent\"])(\"KyoTitlebarSearch\")\n  const _component_NButton = Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"resolveComponent\"])(\"NButton\")\n  const _component_KyoTitlebar = Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"resolveComponent\"])(\"KyoTitlebar\")\n  const _component_NTableColumn = Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"resolveComponent\"])(\"NTableColumn\")\n  const _component_router_link = Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"resolveComponent\"])(\"router-link\")\n  const _component_KyoDatatable = Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"resolveComponent\"])(\"KyoDatatable\")\n  const _component_NLoader = Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"resolveComponent\"])(\"NLoader\")\n\n  return (Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"openBlock\"])(), Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createBlock\"])(_component_NLoader, {\n    visible: _ctx.load,\n    class: \"full-height-child\"\n  }, {\n    default: Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"withCtx\"])(() => [\n      Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createVNode\"])(\"div\", _hoisted_1, [\n        Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createVNode\"])(_component_KyoTitlebar, {\n          class: \"col--flex-0-0\",\n          onDelete: _ctx.deleteItems\n        }, {\n          search: Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"withCtx\"])(() => [\n            Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createVNode\"])(_component_KyoTitlebarSearch, {\n              modelValue: _ctx.query.search,\n              \"onUpdate:modelValue\": _cache[1] || (_cache[1] = $event => (_ctx.query.search = $event))\n            }, null, 8 /* PROPS */, [\"modelValue\"])\n          ]),\n          action: Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"withCtx\"])(() => [\n            Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createVNode\"])(_component_NButton, {\n              type: \"primary\",\n              onClick: _cache[2] || (_cache[2] = $event => (_ctx.$router.push({ name: 'KyoPageCreate' })))\n            }, {\n              default: Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"withCtx\"])(() => [\n                Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createTextVNode\"])(Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"toDisplayString\"])(_ctx.trans('Create page')), 1 /* TEXT */)\n              ]),\n              _: 1 /* STABLE */\n            })\n          ]),\n          _: 1 /* STABLE */\n        }, 8 /* PROPS */, [\"onDelete\"]),\n        Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createVNode\"])(_component_KyoDatatable, {\n          class: \"col--flex-1-1\",\n          onRowDblclick: _ctx.gotoEdit,\n          group: ['page', 'transaction']\n        }, {\n          default: Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"withCtx\"])(() => [\n            Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createVNode\"])(_component_NTableColumn, {\n              type: \"boolean\",\n              prop: \"state\",\n              label: _ctx.trans('State'),\n              sort: true,\n              filter: true\n            }, {\n              default: Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"withCtx\"])(() => [\n                Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createCommentVNode\"])(\" Column \")\n              ]),\n              _: 1 /* STABLE */\n            }, 8 /* PROPS */, [\"label\"]),\n            Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createVNode\"])(_component_NTableColumn, {\n              type: \"boolean\",\n              prop: \"hide\",\n              label: _ctx.trans('Hide'),\n              sort: true,\n              filter: true\n            }, {\n              default: Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"withCtx\"])(() => [\n                Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createCommentVNode\"])(\" Column \")\n              ]),\n              _: 1 /* STABLE */\n            }, 8 /* PROPS */, [\"label\"]),\n            Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createVNode\"])(_component_NTableColumn, {\n              type: \"string\",\n              prop: \"title\",\n              label: _ctx.trans('Title'),\n              fluid: true,\n              sort: true,\n              filter: true\n            }, {\n              default: Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"withCtx\"])(({ item }) => [\n                Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createVNode\"])(_component_router_link, {\n                  to: { name: 'KyoPageEdit', params: item }\n                }, {\n                  default: Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"withCtx\"])(() => [\n                    Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createTextVNode\"])(Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"toDisplayString\"])(item.title), 1 /* TEXT */)\n                  ]),\n                  _: 2 /* DYNAMIC */\n                }, 1032 /* PROPS, DYNAMIC_SLOTS */, [\"to\"])\n              ]),\n              _: 1 /* STABLE */\n            }, 8 /* PROPS */, [\"label\"]),\n            Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createVNode\"])(_component_NTableColumn, {\n              type: \"datetime\",\n              prop: \"updated_at\",\n              label: _ctx.trans('Modified'),\n              sort: true,\n              filter: true\n            }, {\n              default: Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"withCtx\"])(() => [\n                Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createCommentVNode\"])(\" Column \")\n              ]),\n              _: 1 /* STABLE */\n            }, 8 /* PROPS */, [\"label\"]),\n            Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createVNode\"])(_component_NTableColumn, {\n              type: \"datetime\",\n              prop: \"created_at\",\n              label: _ctx.trans('Created'),\n              sort: true,\n              filter: true\n            }, {\n              default: Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"withCtx\"])(() => [\n                Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createCommentVNode\"])(\" Column \")\n              ]),\n              _: 1 /* STABLE */\n            }, 8 /* PROPS */, [\"label\"])\n          ]),\n          _: 1 /* STABLE */\n        }, 8 /* PROPS */, [\"onRowDblclick\"])\n      ])\n    ]),\n    _: 1 /* STABLE */\n  }, 8 /* PROPS */, [\"visible\"]))\n}\n\n//# sourceURL=webpack:///./resources/js/pages/KyoPages.vue?./node_modules/vue-loader/dist/templateLoader.js??ref--5!./node_modules/vue-loader/dist??ref--8-0");
+
+/***/ }),
+
+/***/ "./node_modules/vue-loader/dist/templateLoader.js?!./node_modules/vue-loader/dist/index.js?!./resources/js/widgets/KyoWidgetHeadline.vue?vue&type=template&id=7eeb1a6a":
+/*!************************************************************************************************************************************************************************************!*\
+  !*** ./node_modules/vue-loader/dist/templateLoader.js??ref--5!./node_modules/vue-loader/dist??ref--8-0!./resources/js/widgets/KyoWidgetHeadline.vue?vue&type=template&id=7eeb1a6a ***!
+  \************************************************************************************************************************************************************************************/
+/*! exports provided: render */
+/***/ (function(module, __webpack_exports__, __webpack_require__) {
+
+"use strict";
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, \"render\", function() { return render; });\n/* harmony import */ var vue__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! vue */ \"vue\");\n/* harmony import */ var vue__WEBPACK_IMPORTED_MODULE_0___default = /*#__PURE__*/__webpack_require__.n(vue__WEBPACK_IMPORTED_MODULE_0__);\n\n\nfunction render(_ctx, _cache, $props, $setup, $data, $options) {\n  const _component_NInput = Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"resolveComponent\"])(\"NInput\")\n  const _component_NFormItem = Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"resolveComponent\"])(\"NFormItem\")\n\n  return (Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"openBlock\"])(), Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createBlock\"])(\"div\", null, [\n    Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createVNode\"])(_component_NFormItem, {\n      label: _ctx.trans('Title')\n    }, {\n      default: Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"withCtx\"])(() => [\n        Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createVNode\"])(_component_NInput, {\n          modelValue: $options.value.title,\n          \"onUpdate:modelValue\": _cache[1] || (_cache[1] = $event => ($options.value.title = $event))\n        }, null, 8 /* PROPS */, [\"modelValue\"])\n      ]),\n      _: 1 /* STABLE */\n    }, 8 /* PROPS */, [\"label\"])\n  ]))\n}\n\n//# sourceURL=webpack:///./resources/js/widgets/KyoWidgetHeadline.vue?./node_modules/vue-loader/dist/templateLoader.js??ref--5!./node_modules/vue-loader/dist??ref--8-0");
 
 /***/ }),
 
@@ -118,43 +190,163 @@ eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export (binding) *
 /***/ (function(module, __webpack_exports__, __webpack_require__) {
 
 "use strict";
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var _pages_KyoPages2__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./pages/KyoPages2 */ \"./resources/js/pages/KyoPages2.vue\");\n// import KyoPageCreate from \"./pages/KyoPageCreate\";\n// App.component(KyoPageCreate.name, KyoPageCreate);\n// import KyoPageEdit from \"./pages/KyoPageEdit\";\n// App.component(KyoPageEdit.name, KyoPageEdit);\n// import KyoPageBuilder from \"./builder/KyoPageBuilder\";\n// App.component(KyoPageBuilder.name, KyoPageBuilder);\n// import KyoWidgetHeadline from \"./widgets/KyoWidgetHeadline\";\n// App.component(KyoWidgetHeadline.name, KyoWidgetHeadline);\n// require('./configs/page');\n\nKyoto.ready(function (App) {\n  App.component(_pages_KyoPages2__WEBPACK_IMPORTED_MODULE_0__[\"default\"].name, _pages_KyoPages2__WEBPACK_IMPORTED_MODULE_0__[\"default\"]);\n});\n\nif (console && console.log) {\n  console.log('kyoto/page ready.');\n}\n\n//# sourceURL=webpack:///./resources/js/bootstrap.js?");
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var _pages_KyoPages__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./pages/KyoPages */ \"./resources/js/pages/KyoPages.vue\");\n/* harmony import */ var _pages_KyoPageCreate__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ./pages/KyoPageCreate */ \"./resources/js/pages/KyoPageCreate.vue\");\n/* harmony import */ var _pages_KyoPageEdit__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ./pages/KyoPageEdit */ \"./resources/js/pages/KyoPageEdit.vue\");\n/* harmony import */ var _builder_KyoPageBuilder__WEBPACK_IMPORTED_MODULE_3__ = __webpack_require__(/*! ./builder/KyoPageBuilder */ \"./resources/js/builder/KyoPageBuilder.jsx\");\n/* harmony import */ var _widgets_KyoWidgetHeadline__WEBPACK_IMPORTED_MODULE_4__ = __webpack_require__(/*! ./widgets/KyoWidgetHeadline */ \"./resources/js/widgets/KyoWidgetHeadline.vue\");\n\nKyoto.component(_pages_KyoPages__WEBPACK_IMPORTED_MODULE_0__[\"default\"].name, _pages_KyoPages__WEBPACK_IMPORTED_MODULE_0__[\"default\"]);\n\nKyoto.component(_pages_KyoPageCreate__WEBPACK_IMPORTED_MODULE_1__[\"default\"].name, _pages_KyoPageCreate__WEBPACK_IMPORTED_MODULE_1__[\"default\"]);\n\nKyoto.component(_pages_KyoPageEdit__WEBPACK_IMPORTED_MODULE_2__[\"default\"].name, _pages_KyoPageEdit__WEBPACK_IMPORTED_MODULE_2__[\"default\"]);\n\nKyoto.component(_builder_KyoPageBuilder__WEBPACK_IMPORTED_MODULE_3__[\"default\"].name, _builder_KyoPageBuilder__WEBPACK_IMPORTED_MODULE_3__[\"default\"]);\n\nKyoto.component(_widgets_KyoWidgetHeadline__WEBPACK_IMPORTED_MODULE_4__[\"default\"].name, _widgets_KyoWidgetHeadline__WEBPACK_IMPORTED_MODULE_4__[\"default\"]);\n\nif (console && console.log) {\n  console.log('kyoto/page ready.');\n}\n\n//# sourceURL=webpack:///./resources/js/bootstrap.js?");
 
 /***/ }),
 
-/***/ "./resources/js/pages/KyoPages2.vue":
-/*!******************************************!*\
-  !*** ./resources/js/pages/KyoPages2.vue ***!
-  \******************************************/
+/***/ "./resources/js/builder/KyoPageBuilder.jsx":
+/*!*************************************************!*\
+  !*** ./resources/js/builder/KyoPageBuilder.jsx ***!
+  \*************************************************/
 /*! exports provided: default */
 /***/ (function(module, __webpack_exports__, __webpack_require__) {
 
 "use strict";
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var _KyoPages2_vue_vue_type_template_id_7c630b54__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./KyoPages2.vue?vue&type=template&id=7c630b54 */ \"./resources/js/pages/KyoPages2.vue?vue&type=template&id=7c630b54\");\n/* harmony import */ var _KyoPages2_vue_vue_type_script_lang_js__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ./KyoPages2.vue?vue&type=script&lang=js */ \"./resources/js/pages/KyoPages2.vue?vue&type=script&lang=js\");\n/* empty/unused harmony star reexport */\n\n\n_KyoPages2_vue_vue_type_script_lang_js__WEBPACK_IMPORTED_MODULE_1__[\"default\"].render = _KyoPages2_vue_vue_type_template_id_7c630b54__WEBPACK_IMPORTED_MODULE_0__[\"render\"]\n/* hot reload */\nif (false) {}\n\n_KyoPages2_vue_vue_type_script_lang_js__WEBPACK_IMPORTED_MODULE_1__[\"default\"].__file = \"resources/js/pages/KyoPages2.vue\"\n\n/* harmony default export */ __webpack_exports__[\"default\"] = (_KyoPages2_vue_vue_type_script_lang_js__WEBPACK_IMPORTED_MODULE_1__[\"default\"]);\n\n//# sourceURL=webpack:///./resources/js/pages/KyoPages2.vue?");
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var vue__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! vue */ \"vue\");\n/* harmony import */ var vue__WEBPACK_IMPORTED_MODULE_0___default = /*#__PURE__*/__webpack_require__.n(vue__WEBPACK_IMPORTED_MODULE_0__);\n\n/* harmony default export */ __webpack_exports__[\"default\"] = ({\n  name: 'KyoPageWidgets',\n  props: {\n    value: {\n      \"default\": function _default() {\n        return [];\n      }\n    }\n  },\n  data: function data() {\n    var widgets = [{\n      id: Nano.UUID(),\n      value: {\n        title: ''\n      },\n      type: 'headline',\n      name: 'test1'\n    }, {\n      id: Nano.UUID(),\n      value: {\n        title: ''\n      },\n      type: 'headline',\n      name: 'test2'\n    }];\n    return {\n      widgets: widgets\n    };\n  },\n  methods: {\n    transformDrop: function transformDrop(source) {\n      return Nano.Obj.assign(source, {\n        id: Nano.UUID()\n      });\n    },\n    allowDrop: function allowDrop(source, target, strategy) {\n      return !this.value.length || strategy !== 'root';\n    }\n  },\n  renderWidgetNode: function renderWidgetNode(props) {\n    return Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createVNode\"])(\"div\", {\n      \"class\": \"kyo-widgets-widget-item\"\n    }, [props.item.type]);\n  },\n  renderBuilderNode: function renderBuilderNode(props) {\n    var _this = this;\n\n    return Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createVNode\"])(Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"resolveComponent\"])(\"NForm\"), {\n      \"class\": \"kyo-widgets-builder-item\"\n    }, {\n      \"default\": function _default() {\n        return [Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createVNode\"])(\"div\", {\n          \"class\": \"kyo-widgets-builder-item__head\",\n          \"draggable\": \"true\"\n        }, [Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createVNode\"])(\"div\", {\n          \"class\": \"grid grid--row grid--middle grid--10\"\n        }, [Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createVNode\"])(\"div\", {\n          \"class\": \"col--auto col--left\"\n        }, [props.item.type]), Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createVNode\"])(\"div\", {\n          \"class\": \"col--auto\"\n        }, [Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"withDirectives\"])(Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createVNode\"])(Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"resolveComponent\"])(\"NButton\"), {\n          \"size\": \"mini\",\n          \"type\": \"info\",\n          \"square\": true,\n          \"icon\": _this.icons.create\n        }, null), [[Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"resolveDirective\"])(\"on:click\"), props.copy]])]), Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createVNode\"])(\"div\", {\n          \"class\": \"col--auto\"\n        }, [Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"withDirectives\"])(Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createVNode\"])(Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"resolveComponent\"])(\"NButton\"), {\n          \"size\": \"mini\",\n          \"type\": \"danger\",\n          \"square\": true,\n          \"icon\": _this.icons.times\n        }, null), [[Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"resolveDirective\"])(\"on:click\"), props.remove]])])])]), Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createVNode\"])(\"div\", {\n          \"class\": \"kyo-widgets-builder-item__body\"\n        }, [Vue.h(Vue.resolveComponent(\"KyoWidget\".concat(Nano.Str.ucfirst(props.item.type))))])];\n      }\n    });\n  },\n  render: function render() {\n    var _this2 = this;\n\n    var builderProps = {\n      items: this.value,\n      group: ['page-widget'],\n      allowGroups: ['page-widget'],\n      handle: true,\n      itemHeight: 0,\n      viewportHeight: true,\n      removeNode: false,\n      insertNode: true,\n      useRenderCache: false,\n      keyEvents: false,\n      threshold: 0,\n      allowCurrent: false,\n      transformDrop: this.transformDrop,\n      allowDrop: this.allowDrop,\n      renderNode: this.ctor('renderBuilderNode')\n    };\n    var widgetProps = {\n      items: this.widgets,\n      group: ['page-widget'],\n      itemHeight: 0,\n      viewportHeight: true,\n      removeNode: false,\n      insertNode: false,\n      useRenderCache: false,\n      keyEvents: false,\n      threshold: 0,\n      allowCurrent: false,\n      allowDrop: false,\n      renderNode: this.ctor('renderWidgetNode')\n    };\n    var builderEvents = {\n      'input': function input(value) {\n        return _this2.$emit('input', value);\n      }\n    };\n    return Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createVNode\"])(\"div\", {\n      \"class\": \"kyo-widgets\"\n    }, [Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createVNode\"])(\"div\", {\n      \"class\": \"kyo-widgets__inner\"\n    }, [Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createVNode\"])(Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"resolveComponent\"])(\"NDraglist\"), Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"mergeProps\"])({\n      \"class\": \"kyo-widgets__widget\",\n      \"ref\": \"widget\"\n    }, widgetProps), null), Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createVNode\"])(Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"resolveComponent\"])(\"NDraglist\"), Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"mergeProps\"])({\n      \"class\": \"kyo-widgets__builder\",\n      \"ref\": \"builder\"\n    }, builderProps), null), Object(vue__WEBPACK_IMPORTED_MODULE_0__[\"createVNode\"])(\"div\", {\n      \"class\": \"kyo-widgets__extra\"\n    }, [this.$slots[\"default\"] && this.$slots[\"default\"]()])])]);\n  }\n});\n\n//# sourceURL=webpack:///./resources/js/builder/KyoPageBuilder.jsx?");
 
 /***/ }),
 
-/***/ "./resources/js/pages/KyoPages2.vue?vue&type=script&lang=js":
-/*!******************************************************************!*\
-  !*** ./resources/js/pages/KyoPages2.vue?vue&type=script&lang=js ***!
-  \******************************************************************/
+/***/ "./resources/js/pages/KyoPageCreate.vue":
+/*!**********************************************!*\
+  !*** ./resources/js/pages/KyoPageCreate.vue ***!
+  \**********************************************/
 /*! exports provided: default */
 /***/ (function(module, __webpack_exports__, __webpack_require__) {
 
 "use strict";
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var _node_modules_babel_loader_lib_index_js_ref_0_node_modules_vue_loader_dist_index_js_ref_8_0_KyoPages2_vue_vue_type_script_lang_js__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! -!../../../node_modules/babel-loader/lib??ref--0!../../../node_modules/vue-loader/dist??ref--8-0!./KyoPages2.vue?vue&type=script&lang=js */ \"./node_modules/babel-loader/lib/index.js?!./node_modules/vue-loader/dist/index.js?!./resources/js/pages/KyoPages2.vue?vue&type=script&lang=js\");\n/* harmony reexport (safe) */ __webpack_require__.d(__webpack_exports__, \"default\", function() { return _node_modules_babel_loader_lib_index_js_ref_0_node_modules_vue_loader_dist_index_js_ref_8_0_KyoPages2_vue_vue_type_script_lang_js__WEBPACK_IMPORTED_MODULE_0__[\"default\"]; });\n\n/* empty/unused harmony star reexport */ \n\n//# sourceURL=webpack:///./resources/js/pages/KyoPages2.vue?");
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var _KyoPageCreate_vue_vue_type_template_id_22f38b53__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./KyoPageCreate.vue?vue&type=template&id=22f38b53 */ \"./resources/js/pages/KyoPageCreate.vue?vue&type=template&id=22f38b53\");\n/* harmony import */ var _KyoPageCreate_vue_vue_type_script_lang_js__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ./KyoPageCreate.vue?vue&type=script&lang=js */ \"./resources/js/pages/KyoPageCreate.vue?vue&type=script&lang=js\");\n/* empty/unused harmony star reexport */\n\n\n_KyoPageCreate_vue_vue_type_script_lang_js__WEBPACK_IMPORTED_MODULE_1__[\"default\"].render = _KyoPageCreate_vue_vue_type_template_id_22f38b53__WEBPACK_IMPORTED_MODULE_0__[\"render\"]\n/* hot reload */\nif (false) {}\n\n_KyoPageCreate_vue_vue_type_script_lang_js__WEBPACK_IMPORTED_MODULE_1__[\"default\"].__file = \"resources/js/pages/KyoPageCreate.vue\"\n\n/* harmony default export */ __webpack_exports__[\"default\"] = (_KyoPageCreate_vue_vue_type_script_lang_js__WEBPACK_IMPORTED_MODULE_1__[\"default\"]);\n\n//# sourceURL=webpack:///./resources/js/pages/KyoPageCreate.vue?");
 
 /***/ }),
 
-/***/ "./resources/js/pages/KyoPages2.vue?vue&type=template&id=7c630b54":
-/*!************************************************************************!*\
-  !*** ./resources/js/pages/KyoPages2.vue?vue&type=template&id=7c630b54 ***!
-  \************************************************************************/
+/***/ "./resources/js/pages/KyoPageCreate.vue?vue&type=script&lang=js":
+/*!**********************************************************************!*\
+  !*** ./resources/js/pages/KyoPageCreate.vue?vue&type=script&lang=js ***!
+  \**********************************************************************/
+/*! exports provided: default */
+/***/ (function(module, __webpack_exports__, __webpack_require__) {
+
+"use strict";
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var _node_modules_babel_loader_lib_index_js_ref_0_node_modules_vue_loader_dist_index_js_ref_8_0_KyoPageCreate_vue_vue_type_script_lang_js__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! -!../../../node_modules/babel-loader/lib??ref--0!../../../node_modules/vue-loader/dist??ref--8-0!./KyoPageCreate.vue?vue&type=script&lang=js */ \"./node_modules/babel-loader/lib/index.js?!./node_modules/vue-loader/dist/index.js?!./resources/js/pages/KyoPageCreate.vue?vue&type=script&lang=js\");\n/* harmony reexport (safe) */ __webpack_require__.d(__webpack_exports__, \"default\", function() { return _node_modules_babel_loader_lib_index_js_ref_0_node_modules_vue_loader_dist_index_js_ref_8_0_KyoPageCreate_vue_vue_type_script_lang_js__WEBPACK_IMPORTED_MODULE_0__[\"default\"]; });\n\n/* empty/unused harmony star reexport */ \n\n//# sourceURL=webpack:///./resources/js/pages/KyoPageCreate.vue?");
+
+/***/ }),
+
+/***/ "./resources/js/pages/KyoPageCreate.vue?vue&type=template&id=22f38b53":
+/*!****************************************************************************!*\
+  !*** ./resources/js/pages/KyoPageCreate.vue?vue&type=template&id=22f38b53 ***!
+  \****************************************************************************/
 /*! exports provided: render */
 /***/ (function(module, __webpack_exports__, __webpack_require__) {
 
 "use strict";
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var _node_modules_vue_loader_dist_templateLoader_js_ref_5_node_modules_vue_loader_dist_index_js_ref_8_0_KyoPages2_vue_vue_type_template_id_7c630b54__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! -!../../../node_modules/vue-loader/dist/templateLoader.js??ref--5!../../../node_modules/vue-loader/dist??ref--8-0!./KyoPages2.vue?vue&type=template&id=7c630b54 */ \"./node_modules/vue-loader/dist/templateLoader.js?!./node_modules/vue-loader/dist/index.js?!./resources/js/pages/KyoPages2.vue?vue&type=template&id=7c630b54\");\n/* harmony reexport (safe) */ __webpack_require__.d(__webpack_exports__, \"render\", function() { return _node_modules_vue_loader_dist_templateLoader_js_ref_5_node_modules_vue_loader_dist_index_js_ref_8_0_KyoPages2_vue_vue_type_template_id_7c630b54__WEBPACK_IMPORTED_MODULE_0__[\"render\"]; });\n\n\n\n//# sourceURL=webpack:///./resources/js/pages/KyoPages2.vue?");
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var _node_modules_vue_loader_dist_templateLoader_js_ref_5_node_modules_vue_loader_dist_index_js_ref_8_0_KyoPageCreate_vue_vue_type_template_id_22f38b53__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! -!../../../node_modules/vue-loader/dist/templateLoader.js??ref--5!../../../node_modules/vue-loader/dist??ref--8-0!./KyoPageCreate.vue?vue&type=template&id=22f38b53 */ \"./node_modules/vue-loader/dist/templateLoader.js?!./node_modules/vue-loader/dist/index.js?!./resources/js/pages/KyoPageCreate.vue?vue&type=template&id=22f38b53\");\n/* harmony reexport (safe) */ __webpack_require__.d(__webpack_exports__, \"render\", function() { return _node_modules_vue_loader_dist_templateLoader_js_ref_5_node_modules_vue_loader_dist_index_js_ref_8_0_KyoPageCreate_vue_vue_type_template_id_22f38b53__WEBPACK_IMPORTED_MODULE_0__[\"render\"]; });\n\n\n\n//# sourceURL=webpack:///./resources/js/pages/KyoPageCreate.vue?");
+
+/***/ }),
+
+/***/ "./resources/js/pages/KyoPageEdit.vue":
+/*!********************************************!*\
+  !*** ./resources/js/pages/KyoPageEdit.vue ***!
+  \********************************************/
+/*! exports provided: default */
+/***/ (function(module, __webpack_exports__, __webpack_require__) {
+
+"use strict";
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var _KyoPageEdit_vue_vue_type_template_id_d1e6f5be__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./KyoPageEdit.vue?vue&type=template&id=d1e6f5be */ \"./resources/js/pages/KyoPageEdit.vue?vue&type=template&id=d1e6f5be\");\n/* harmony import */ var _KyoPageEdit_vue_vue_type_script_lang_js__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ./KyoPageEdit.vue?vue&type=script&lang=js */ \"./resources/js/pages/KyoPageEdit.vue?vue&type=script&lang=js\");\n/* empty/unused harmony star reexport */\n\n\n_KyoPageEdit_vue_vue_type_script_lang_js__WEBPACK_IMPORTED_MODULE_1__[\"default\"].render = _KyoPageEdit_vue_vue_type_template_id_d1e6f5be__WEBPACK_IMPORTED_MODULE_0__[\"render\"]\n/* hot reload */\nif (false) {}\n\n_KyoPageEdit_vue_vue_type_script_lang_js__WEBPACK_IMPORTED_MODULE_1__[\"default\"].__file = \"resources/js/pages/KyoPageEdit.vue\"\n\n/* harmony default export */ __webpack_exports__[\"default\"] = (_KyoPageEdit_vue_vue_type_script_lang_js__WEBPACK_IMPORTED_MODULE_1__[\"default\"]);\n\n//# sourceURL=webpack:///./resources/js/pages/KyoPageEdit.vue?");
+
+/***/ }),
+
+/***/ "./resources/js/pages/KyoPageEdit.vue?vue&type=script&lang=js":
+/*!********************************************************************!*\
+  !*** ./resources/js/pages/KyoPageEdit.vue?vue&type=script&lang=js ***!
+  \********************************************************************/
+/*! exports provided: default */
+/***/ (function(module, __webpack_exports__, __webpack_require__) {
+
+"use strict";
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var _node_modules_babel_loader_lib_index_js_ref_0_node_modules_vue_loader_dist_index_js_ref_8_0_KyoPageEdit_vue_vue_type_script_lang_js__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! -!../../../node_modules/babel-loader/lib??ref--0!../../../node_modules/vue-loader/dist??ref--8-0!./KyoPageEdit.vue?vue&type=script&lang=js */ \"./node_modules/babel-loader/lib/index.js?!./node_modules/vue-loader/dist/index.js?!./resources/js/pages/KyoPageEdit.vue?vue&type=script&lang=js\");\n/* harmony reexport (safe) */ __webpack_require__.d(__webpack_exports__, \"default\", function() { return _node_modules_babel_loader_lib_index_js_ref_0_node_modules_vue_loader_dist_index_js_ref_8_0_KyoPageEdit_vue_vue_type_script_lang_js__WEBPACK_IMPORTED_MODULE_0__[\"default\"]; });\n\n/* empty/unused harmony star reexport */ \n\n//# sourceURL=webpack:///./resources/js/pages/KyoPageEdit.vue?");
+
+/***/ }),
+
+/***/ "./resources/js/pages/KyoPageEdit.vue?vue&type=template&id=d1e6f5be":
+/*!**************************************************************************!*\
+  !*** ./resources/js/pages/KyoPageEdit.vue?vue&type=template&id=d1e6f5be ***!
+  \**************************************************************************/
+/*! exports provided: render */
+/***/ (function(module, __webpack_exports__, __webpack_require__) {
+
+"use strict";
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var _node_modules_vue_loader_dist_templateLoader_js_ref_5_node_modules_vue_loader_dist_index_js_ref_8_0_KyoPageEdit_vue_vue_type_template_id_d1e6f5be__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! -!../../../node_modules/vue-loader/dist/templateLoader.js??ref--5!../../../node_modules/vue-loader/dist??ref--8-0!./KyoPageEdit.vue?vue&type=template&id=d1e6f5be */ \"./node_modules/vue-loader/dist/templateLoader.js?!./node_modules/vue-loader/dist/index.js?!./resources/js/pages/KyoPageEdit.vue?vue&type=template&id=d1e6f5be\");\n/* harmony reexport (safe) */ __webpack_require__.d(__webpack_exports__, \"render\", function() { return _node_modules_vue_loader_dist_templateLoader_js_ref_5_node_modules_vue_loader_dist_index_js_ref_8_0_KyoPageEdit_vue_vue_type_template_id_d1e6f5be__WEBPACK_IMPORTED_MODULE_0__[\"render\"]; });\n\n\n\n//# sourceURL=webpack:///./resources/js/pages/KyoPageEdit.vue?");
+
+/***/ }),
+
+/***/ "./resources/js/pages/KyoPages.vue":
+/*!*****************************************!*\
+  !*** ./resources/js/pages/KyoPages.vue ***!
+  \*****************************************/
+/*! exports provided: default */
+/***/ (function(module, __webpack_exports__, __webpack_require__) {
+
+"use strict";
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var _KyoPages_vue_vue_type_template_id_671d7ee8__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./KyoPages.vue?vue&type=template&id=671d7ee8 */ \"./resources/js/pages/KyoPages.vue?vue&type=template&id=671d7ee8\");\n/* harmony import */ var _KyoPages_vue_vue_type_script_lang_js__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ./KyoPages.vue?vue&type=script&lang=js */ \"./resources/js/pages/KyoPages.vue?vue&type=script&lang=js\");\n/* empty/unused harmony star reexport */\n\n\n_KyoPages_vue_vue_type_script_lang_js__WEBPACK_IMPORTED_MODULE_1__[\"default\"].render = _KyoPages_vue_vue_type_template_id_671d7ee8__WEBPACK_IMPORTED_MODULE_0__[\"render\"]\n/* hot reload */\nif (false) {}\n\n_KyoPages_vue_vue_type_script_lang_js__WEBPACK_IMPORTED_MODULE_1__[\"default\"].__file = \"resources/js/pages/KyoPages.vue\"\n\n/* harmony default export */ __webpack_exports__[\"default\"] = (_KyoPages_vue_vue_type_script_lang_js__WEBPACK_IMPORTED_MODULE_1__[\"default\"]);\n\n//# sourceURL=webpack:///./resources/js/pages/KyoPages.vue?");
+
+/***/ }),
+
+/***/ "./resources/js/pages/KyoPages.vue?vue&type=script&lang=js":
+/*!*****************************************************************!*\
+  !*** ./resources/js/pages/KyoPages.vue?vue&type=script&lang=js ***!
+  \*****************************************************************/
+/*! exports provided: default */
+/***/ (function(module, __webpack_exports__, __webpack_require__) {
+
+"use strict";
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var _node_modules_babel_loader_lib_index_js_ref_0_node_modules_vue_loader_dist_index_js_ref_8_0_KyoPages_vue_vue_type_script_lang_js__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! -!../../../node_modules/babel-loader/lib??ref--0!../../../node_modules/vue-loader/dist??ref--8-0!./KyoPages.vue?vue&type=script&lang=js */ \"./node_modules/babel-loader/lib/index.js?!./node_modules/vue-loader/dist/index.js?!./resources/js/pages/KyoPages.vue?vue&type=script&lang=js\");\n/* harmony reexport (safe) */ __webpack_require__.d(__webpack_exports__, \"default\", function() { return _node_modules_babel_loader_lib_index_js_ref_0_node_modules_vue_loader_dist_index_js_ref_8_0_KyoPages_vue_vue_type_script_lang_js__WEBPACK_IMPORTED_MODULE_0__[\"default\"]; });\n\n/* empty/unused harmony star reexport */ \n\n//# sourceURL=webpack:///./resources/js/pages/KyoPages.vue?");
+
+/***/ }),
+
+/***/ "./resources/js/pages/KyoPages.vue?vue&type=template&id=671d7ee8":
+/*!***********************************************************************!*\
+  !*** ./resources/js/pages/KyoPages.vue?vue&type=template&id=671d7ee8 ***!
+  \***********************************************************************/
+/*! exports provided: render */
+/***/ (function(module, __webpack_exports__, __webpack_require__) {
+
+"use strict";
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var _node_modules_vue_loader_dist_templateLoader_js_ref_5_node_modules_vue_loader_dist_index_js_ref_8_0_KyoPages_vue_vue_type_template_id_671d7ee8__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! -!../../../node_modules/vue-loader/dist/templateLoader.js??ref--5!../../../node_modules/vue-loader/dist??ref--8-0!./KyoPages.vue?vue&type=template&id=671d7ee8 */ \"./node_modules/vue-loader/dist/templateLoader.js?!./node_modules/vue-loader/dist/index.js?!./resources/js/pages/KyoPages.vue?vue&type=template&id=671d7ee8\");\n/* harmony reexport (safe) */ __webpack_require__.d(__webpack_exports__, \"render\", function() { return _node_modules_vue_loader_dist_templateLoader_js_ref_5_node_modules_vue_loader_dist_index_js_ref_8_0_KyoPages_vue_vue_type_template_id_671d7ee8__WEBPACK_IMPORTED_MODULE_0__[\"render\"]; });\n\n\n\n//# sourceURL=webpack:///./resources/js/pages/KyoPages.vue?");
+
+/***/ }),
+
+/***/ "./resources/js/widgets/KyoWidgetHeadline.vue":
+/*!****************************************************!*\
+  !*** ./resources/js/widgets/KyoWidgetHeadline.vue ***!
+  \****************************************************/
+/*! exports provided: default */
+/***/ (function(module, __webpack_exports__, __webpack_require__) {
+
+"use strict";
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var _KyoWidgetHeadline_vue_vue_type_template_id_7eeb1a6a__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./KyoWidgetHeadline.vue?vue&type=template&id=7eeb1a6a */ \"./resources/js/widgets/KyoWidgetHeadline.vue?vue&type=template&id=7eeb1a6a\");\n/* harmony import */ var _KyoWidgetHeadline_vue_vue_type_script_lang_js__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ./KyoWidgetHeadline.vue?vue&type=script&lang=js */ \"./resources/js/widgets/KyoWidgetHeadline.vue?vue&type=script&lang=js\");\n/* empty/unused harmony star reexport */\n\n\n_KyoWidgetHeadline_vue_vue_type_script_lang_js__WEBPACK_IMPORTED_MODULE_1__[\"default\"].render = _KyoWidgetHeadline_vue_vue_type_template_id_7eeb1a6a__WEBPACK_IMPORTED_MODULE_0__[\"render\"]\n/* hot reload */\nif (false) {}\n\n_KyoWidgetHeadline_vue_vue_type_script_lang_js__WEBPACK_IMPORTED_MODULE_1__[\"default\"].__file = \"resources/js/widgets/KyoWidgetHeadline.vue\"\n\n/* harmony default export */ __webpack_exports__[\"default\"] = (_KyoWidgetHeadline_vue_vue_type_script_lang_js__WEBPACK_IMPORTED_MODULE_1__[\"default\"]);\n\n//# sourceURL=webpack:///./resources/js/widgets/KyoWidgetHeadline.vue?");
+
+/***/ }),
+
+/***/ "./resources/js/widgets/KyoWidgetHeadline.vue?vue&type=script&lang=js":
+/*!****************************************************************************!*\
+  !*** ./resources/js/widgets/KyoWidgetHeadline.vue?vue&type=script&lang=js ***!
+  \****************************************************************************/
+/*! exports provided: default */
+/***/ (function(module, __webpack_exports__, __webpack_require__) {
+
+"use strict";
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var _node_modules_babel_loader_lib_index_js_ref_0_node_modules_vue_loader_dist_index_js_ref_8_0_KyoWidgetHeadline_vue_vue_type_script_lang_js__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! -!../../../node_modules/babel-loader/lib??ref--0!../../../node_modules/vue-loader/dist??ref--8-0!./KyoWidgetHeadline.vue?vue&type=script&lang=js */ \"./node_modules/babel-loader/lib/index.js?!./node_modules/vue-loader/dist/index.js?!./resources/js/widgets/KyoWidgetHeadline.vue?vue&type=script&lang=js\");\n/* harmony reexport (safe) */ __webpack_require__.d(__webpack_exports__, \"default\", function() { return _node_modules_babel_loader_lib_index_js_ref_0_node_modules_vue_loader_dist_index_js_ref_8_0_KyoWidgetHeadline_vue_vue_type_script_lang_js__WEBPACK_IMPORTED_MODULE_0__[\"default\"]; });\n\n/* empty/unused harmony star reexport */ \n\n//# sourceURL=webpack:///./resources/js/widgets/KyoWidgetHeadline.vue?");
+
+/***/ }),
+
+/***/ "./resources/js/widgets/KyoWidgetHeadline.vue?vue&type=template&id=7eeb1a6a":
+/*!**********************************************************************************!*\
+  !*** ./resources/js/widgets/KyoWidgetHeadline.vue?vue&type=template&id=7eeb1a6a ***!
+  \**********************************************************************************/
+/*! exports provided: render */
+/***/ (function(module, __webpack_exports__, __webpack_require__) {
+
+"use strict";
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var _node_modules_vue_loader_dist_templateLoader_js_ref_5_node_modules_vue_loader_dist_index_js_ref_8_0_KyoWidgetHeadline_vue_vue_type_template_id_7eeb1a6a__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! -!../../../node_modules/vue-loader/dist/templateLoader.js??ref--5!../../../node_modules/vue-loader/dist??ref--8-0!./KyoWidgetHeadline.vue?vue&type=template&id=7eeb1a6a */ \"./node_modules/vue-loader/dist/templateLoader.js?!./node_modules/vue-loader/dist/index.js?!./resources/js/widgets/KyoWidgetHeadline.vue?vue&type=template&id=7eeb1a6a\");\n/* harmony reexport (safe) */ __webpack_require__.d(__webpack_exports__, \"render\", function() { return _node_modules_vue_loader_dist_templateLoader_js_ref_5_node_modules_vue_loader_dist_index_js_ref_8_0_KyoWidgetHeadline_vue_vue_type_template_id_7eeb1a6a__WEBPACK_IMPORTED_MODULE_0__[\"render\"]; });\n\n\n\n//# sourceURL=webpack:///./resources/js/widgets/KyoWidgetHeadline.vue?");
 
 /***/ }),
 
