@@ -1,25 +1,19 @@
 
 
-// import KyoPageCreate from "./pages/KyoPageCreate";
-// App.component(KyoPageCreate.name, KyoPageCreate);
+import KyoPages from "./pages/KyoPages";
+Kyoto.component(KyoPages.name, KyoPages);
 
-// import KyoPageEdit from "./pages/KyoPageEdit";
-// App.component(KyoPageEdit.name, KyoPageEdit);
+import KyoPageCreate from "./pages/KyoPageCreate";
+Kyoto.component(KyoPageCreate.name, KyoPageCreate);
 
-// import KyoPageBuilder from "./builder/KyoPageBuilder";
-// App.component(KyoPageBuilder.name, KyoPageBuilder);
+import KyoPageEdit from "./pages/KyoPageEdit";
+Kyoto.component(KyoPageEdit.name, KyoPageEdit);
 
-// import KyoWidgetHeadline from "./widgets/KyoWidgetHeadline";
-// App.component(KyoWidgetHeadline.name, KyoWidgetHeadline);
+import KyoPageBuilder from "./builder/KyoPageBuilder";
+Kyoto.component(KyoPageBuilder.name, KyoPageBuilder);
 
-// require('./configs/page');
-
-import KyoPages from "./pages/KyoPages2";
-
-Kyoto.ready(function (App) {
-    App.component(KyoPages.name, KyoPages);
-});
-
+import KyoWidgetHeadline from "./widgets/KyoWidgetHeadline";
+Kyoto.component(KyoWidgetHeadline.name, KyoWidgetHeadline);
 
 if ( console && console.log ) {
     console.log('kyoto/page ready.');
