@@ -5,11 +5,11 @@
 
             <KyoTitlebar class="col--flex-0-0" @delete="deleteItems">
 
-                <template slot="search">
+                <template v-slot:search>
                     <KyoTitlebarSearch v-model="query.search"></KyoTitlebarSearch>
                 </template>
 
-                <template slot="action">
+                <template v-slot:action>
                     <NButton type="primary" @click="$router.push({ name: 'KyoGateCreate' })">
                         {{ trans('Create gate') }}
                     </NButton>
@@ -20,20 +20,14 @@
             <KyoDatatable class="col--flex-1-1" :render-expand="true" @row-dblclick="gotoEdit">
 
                 <NTableColumn type="string" prop="title" :label="trans('Title')" :fluid="true" :sort="true" :filter="true">
-                    <router-link slot-scope="{ value }" :to="{ name: 'KyoGateEdit', params: value }">{{ value.title }}</router-link>
+                    <template v-slot:default="{ item }">
+                        <router-link :to="{ name: 'KyoGateEdit', params: item }">{{ item.title }}</router-link>
+                    </template>
                 </NTableColumn>
 
-                <NTableColumn type="string" prop="description" :label="trans('Description')" :fluid="true" :sort="true" :filter="true">
-                    <!-- Column -->
-                </NTableColumn>
-
-                <NTableColumn type="datetime" prop="updated_at" :label="trans('Modified')" :sort="true" :filter="true">
-                    <!-- Column -->
-                </NTableColumn>
-
-                <NTableColumn type="datetime" prop="created_at" :label="trans('Created')" :sort="true" :filter="true">
-                    <!-- Column -->
-                </NTableColumn>
+                <NTableColumn type="string" prop="description" :label="trans('Description')" :fluid="true" :sort="true" :filter="true"></NTableColumn>
+                <NTableColumn type="datetime" prop="updated_at" :label="trans('Modified')" :sort="true" :filter="true"></NTableColumn>
+                <NTableColumn type="datetime" prop="created_at" :label="trans('Created')" :sort="true" :filter="true"></NTableColumn>
 
             </KyoDatatable>
 
