@@ -3,7 +3,7 @@
         <div class="grid grid--col">
 
             <KyoTitlebar :link="{ name: 'KyoUsers' }" class="col--flex-0-0">
-                <template slot="action">
+                <template v-slot:action>
                     <NButtonGroup>
                         <NButton type="primary" @click="storeItem">
                             {{ trans('Apply') }}
@@ -17,7 +17,7 @@
 
             <NForm :form="result" :errors="errors" class="kyo-dataform col--flex-1-0">
 
-                <NFormGroup icon="fa fa-cog" :legend="trans('Settings')">
+                <NFormGroup icon="fa fa-cog" :label="trans('Settings')">
 
                     <div class="grid grid-row grid--wrap grid--30">
 
@@ -40,7 +40,7 @@
                     </div>
                 </NFormGroup>
 
-                <NFormGroup icon="fa fa-user" :legend="trans('User')">
+                <NFormGroup icon="fa fa-user" :label="trans('User')">
                     <div class="grid grid-row grid--wrap grid--30">
 
                         <div class="col--1-1 col--1-2@sm">
@@ -58,7 +58,7 @@
                     </div>
                 </NFormGroup>
 
-                <NFormGroup icon="fa fa-user-shield" :legend="trans('Access')">
+                <NFormGroup icon="fa fa-user-shield" :label="trans('Access')">
                     <div class="grid grid-row grid--wrap grid--30">
 
                         <div class="col--1-1 col--1-2@sm">
@@ -76,7 +76,7 @@
                     </div>
                 </NFormGroup>
 
-                <NFormGroup icon="fa fa-key" :legend="trans('Security')">
+                <NFormGroup icon="fa fa-key" :label="trans('Security')">
                     <div class="grid grid-row grid--wrap grid--30">
 
                         <div class="col--1-1 col--1-2@sm">
