@@ -5,11 +5,11 @@
 
             <KyoTitlebar class="col--flex-0-0" @delete="deleteItems">
 
-                <template slot="search">
+                <template v-slot:search>
                     <KyoTitlebarSearch v-model="query.search"></KyoTitlebarSearch>
                 </template>
 
-                <template slot="action">
+                <template v-slot:action>
                     <NButton type="primary" @click="$router.push({ name: 'KyoUserCreate' })">
                         {{ trans('Create user') }}
                     </NButton>
@@ -20,20 +20,14 @@
             <KyoDatatable class="col--flex-1-1" @row-dblclick="gotoEdit">
 
                 <NTableColumn type="string" prop="name" :label="trans('Name')" :fluid="true" :sort="true" :filter="true">
-                    <router-link slot-scope="{ value }" :to="{ name: 'KyoUserEdit', params: value }">{{ value.name }}</router-link>
+                    <template v-slot:default="{ item }">
+                        <router-link :to="{ name: 'KyoUserEdit', params: item }">{{ item.name }}</router-link>
+                    </template>
                 </NTableColumn>
 
-                <NTableColumn type="string" prop="email" :label="trans('E-Mail')" :fluid="true" :sort="true" :filter="true">
-                    <!-- Column -->
-                </NTableColumn>
-
-                <NTableColumn type="datetime" prop="updated_at" :label="trans('Modified')" :sort="true" :filter="true">
-                    <!-- Column -->
-                </NTableColumn>
-
-                <NTableColumn type="datetime" prop="created_at" :label="trans('Created')" :sort="true" :filter="true">
-                    <!-- Column -->
-                </NTableColumn>
+                <NTableColumn type="string" prop="email" :label="trans('E-Mail')" :fluid="true" :sort="true" :filter="true"></NTableColumn>
+                <NTableColumn type="datetime" prop="updated_at" :label="trans('Modified')" :sort="true" :filter="true"></NTableColumn>
+                <NTableColumn type="datetime" prop="created_at" :label="trans('Created')" :sort="true" :filter="true"></NTableColumn>
 
             </KyoDatatable>
 
